ADI-420: Restrict output of debug LDAP user information to only an amount of characters
At the moment the whole array with LDAP user information will be printed out. This works but leads to large debug logs. It gets even worse when dumping binary data like thumbnailphoto.
This fix restricts the output to 32 characters.
In addition to that, the lookup of empty GUIDs will prevented to suppress invalid output.